+++ c_type_shock/coll_rates_co.cpp
@@ -348,7 +348,7 @@ void co_collisions::get_rate_neutrals(const energy_level &up_lev, const energy_l
 	}
 
 	if (down_rate > MIN_COLLISION_RATE)
-		up_rate = down_rate *exp((low_lev.energy - up_lev.energy)*CM_INVERSE_TO_KELVINS/temp_neutrals) *up_lev.g /low_lev.g;
+		up_rate = down_rate *exp((low_lev.energy - up_lev.energy)*CM_INVERSE_TO_KELVINS/temp_neutrals) *up_lev.g /((double) low_lev.g);
 	else down_rate = 0.;
 }
